simplify card components
