--- frontend/src/components/ToolsSection/FeatureCards/components/FeatureCardGrid/index.ts
@@ -1,1 +1,0 @@
-export { default } from './FeatureCardGrid';
