--- frontend/src/components/ToolsSection/FeatureCards/components/FeatureCardGridContainer/index.ts
@@ -1,1 +1,0 @@
-export { default as FeatureCardGridContainer } from './FeatureCardGridContainer';
