--- frontend/src/components/ToolsSection/FeatureCards/components/index.ts
@@ -1,2 +1,0 @@
-export { default as FeatureCard } from './FeatureCardGrid/FeatureCardGrid';
-export { FeatureCardGridContainer } from './FeatureCardGridContainer';
